requirements.txt
amqp==1.4.9
anyjson==0.3.3
billiard==3.3.0.23
celery==3.1.18
Django==2.2.9
djangorestframework==3.11.0
et-xmlfile==1.0.1
jdcal==1.4.1
kombu==3.0.37
numpy==1.18.5
openpyxl==3.0.3
pandas==1.0.4
python-dateutil==2.8.1
pytz==2020.1
redis==2.10.3
six==1.15.0
sqlparse==0.3.1
